OmniAgent_VR_System/CognitiveEngine/requirements.txt
# File: requirements.txt
# Purpose: Python Dependencies for the Cognitive Engine.
# Key Constraints: pydantic>=2.0 for Strict Data Contracts.
mcp

langgraph
langchain
fastapi
uvicorn
# Data Contract validation - Strict V2 required
pydantic>=2.0
python-dotenv
websockets
PyYAML
langchain-openai
